Extract FAAppNavigation composable to separate file
